libtestutil: set FLUX_REACTOR_SIGCHLD on test server reactor
Problem: The handle provided to the test server implementation by
test_server_create() isn't capable of running subprocesses, since
the FLUX_REACTOR_SIGCHLD flag is not set on the underlying reactor.
It appears that by the time the test server thread is running, it is
too late to call flux_set_reactor(3) (returns EINVAL).

Since the express use case for test_server_create() is to test a single
service in isolation, it should be safe to set FLUX_REACTOR_SIGCHLD
(i.e. the server thread will utilize ev_default_loop())

If in the future this becomes a problem, a flag could be percolated
up to the test_server_create() call.