broker/hello: use idset instead of count internally
Problem: merely capturing a count of peers/descendents
is not enough information for future execution target
liveness eventlog.

Prepare for hello protocol collecting live ranks by
internally using an idset instead of a count in the
hello reduction protocol.

Add accessor hello_get_idset() and use it in broker
to report ranks that have checked in.