sharness: fix bashism in removal of PMI and SLURM env vars
Problem: A bashism crept into the code in flux-sharness.sh which
cleans the test environment of PMI_* and SLURM_* environment variables.
This causes the non-bash shells to error with invalid substitution
when the testsuite is run as a Slurm or Flux job.

Fix the bashism with what (I hope) is a POSIX compliant substitution.

Fixes #2899