sharness: fix value of FLUX_SOURCE_DIR with vpath build
Problem: When the source and build tree of flux-core are different,
and sharness tests are run outside of make check, the path for
FLUX_SOURCE_DIR is set to the same as builddir, causing sharness
tests that use scripts from srcdir to fail.

Use SHARNESS_TEST_SRCDIR as the basis for FLUX_SOURCE_DIR when
srcdir isn't set in the environment to get the correct path.