testsuite: fix race in t2612-job-shell-pty.t
Problem: The last set of tests in t2612-job-shell-pty.t is racy
because a short-lived job can complete before flux-job attach
pty attach is successful, and thus job output is lost. This makes
the test prone to failure, especially in slow environments like
Travis-CI.

Synchronize the test with the test job by waiting in the job for
the test to set a key in the job kvs namespace. In the test, wait
for the `shell.start` event before setting the kvs key, so that
we guarantee the pty has been attached.

Fixes #2903