broker: add module disconnects
Problem: when broker modules are unloaded, they do not
generate disconnect messages.

When a connection is dropped to the connector-local module
(e.g. client dies or otherwise closes the connection),
the connector-local module sends disconnect requests
to all services used by that handle. For example, streaming RPCs
can be canceled.

Add the same functionality to the broker for module unload.
This may come in handy once modules are using streaming RPC
services.  Leverage librouter/disconnect.[ch] which creates a
hash of disconnect messages, added to each time a request
message sent by the module uses a new service.  When the hash
is destroyed, messages are sent using a callback.  Tie hash
destruction to destruction of the module_t by the broker.

Since modules can be "routers", only perform this service for
requests that have a route hop count of 1, which indicates
that they originated from a module, not some client connected
to a module.

Fixes #2911