broker/hello: implement hello.idset RPC
Add an RPC that allows the current idset representing
available brokers to be retrieved.  The request can optionally
include FLUX_MSGFLAG_STREAMING to "watch" idset evolution.

A hello.disconnect request causes a pending idset request to be
discarded.

A hello.cancel request causes a pending idset request to be
canceled and receive ECANCELED.