build: add -no-install to all unit test LDFLAGS
Problem: Under some circumstances (e.g. --with-flux-security), some
unit tests may be linked and run against installed libflux-core.so
instead of the built library. This occurs because unit tests are being
built as if they will be installed, and some combination of link-time
arguments is confsuing the linker into linking with the destination
library instaed of the builddir library.

Instead of trying to rework build arguments to subvert the linker
confusion, this patch adds libtool's -no-install option to all necessary
unit test builds. This option unambiguously tells libtool tha tthe
unit test executable is not meant for installation, and thus prevents
the issue in the problem statement above.

As a side benefit, libtool no longer needs to make wrapper scripts
for unit tests, easing the use of tools like gdb and valgrind on these
executables and decluttering the build directory.

See also:
 https://autotools.io/libtool/wrappers.html

Fixes #2916