docker: start MUNGE in published docker images
Problem: The flux-core docker images published by Travis builds are
inconvenient to use because they are built with --with-flux-security
and thus require munged to be started by hand before jobs can be
submitted.

Replace the current default CMD in the docker images with an
ENTRYPOINT that starts munged and then goes on to run a default CMD of
"/usr/bin/flux start /bin/bash". Since CMD is overridden on the docker
run commandline, the default of flux start can be easily overridden.

Therefore, this should now work

 $ docker run -ti fluxrm/flux-core
 flux@60f289a8515c:~$ flux mini run hostname

as well as this:

 $ docker run -ti fluxrm/flux-core flux start -s 4
 flux@60f289a8515c:~$ flux mini run -N4 -n4 hostname

Fixes #2919