librouter: add usock_conn close callback
Problem: The usock_conn interface only allows a server to get a callback
on error, but not normal close of usock connections. In some cases
(mainly testing), it may be useful to install a callback for every
closed connection.

Add an interface to set an optional close callback for usock_conn objects.