librouter: ensure messages are not dropped on EPIPE
Problem: A small potential for lost messages exists when a usock_client
writes one or messages to an API socket then immediately calls
flux_close() or exits. This can occur if the server attempts to write
a message to the client before reading remaining messages in the
socket buffer.  An EPIPE error is returned, and the server immediately
destroys the connection, leaving the written messages unread.

Fix this issue by handling EPIPE error explicitly in conn_write_cb().
Messages to the client are dropped, but the connection is not
destroyed.  Instead, the connection will be closed after read returns
EOF/ECONNRESET.

Fixes #2931