sharness: add support for checking prereq usage
Problem: There is currently no way to catch tests and test scripts
that make use of programs like `jq` without using a prereq or skip_all
to skip tests when the program is not installed. We could run a test
build without jq installed, but this would require maintaining a
separate docker image for testing, and isn't scalable to multiple
prerequisite programs.

This patch adds support to sharness for wrapping any optional programs
such that an error is generated when they are used outside of a prereq
or a new `skip_all_unless_have()` function (to be used at the top of
scripts where all tests should be skipped if a program is missing)

When TEST_CHECK_PREREQS is set, wrapper scripts for all programs noted
in GLOBAL_PROGRAM_PREREQS are created and inserted first in PATH. The
wrapper script will check to ensure the local or global prereq was
checked before executing the program. Otherwise, the wrapper script
will fail with an error message.

This allows a kind of linter for common prereqs such as jq, which
can be enabled by default during CI tests.

Additionally, this sets prerequisites globally instead of in each
test, reducing tests script code duplication for these common
prereqs.