hwloc: drop the `lstopo` subcommand in preparation for hwloc 2.0+
Problem: hwloc 2.0+ no longer supports custom topologies with
multiple-nodes.  So a 2.0+ version of the `lstopo` command is not
guaranteed to work with a multi-node topology, especially if we have to
manually construct that multi-node topology (since the 2.0+ API provides
no mechanisms for this).

Solution: drop the command entirely since it is not a commonly used subcommand