sharness: fix TEST_CHECK_PREREQS for $jq
Problem: When tests use the fully-qualified path to jq (or other
program prerequisites) via $jq, the TEST_CHECK_PREREQS checks
are subverted, since the wrapper script is not called.

When TEST_CHECK_PREREQS is used, reset $prog to the wrapper script
instead of the original script to ensure the wrapper script is
called.

Fixes #2938