t2300-sched-simple.t: reload resource module as needed
In order to keep the resource module synchronized with the
"fake" by_rank configuration put into the KVS during sched-simple
testing, the resource module needs to be reloaded along with
sched-simple.

This has no present effect, but will be required when sched-simple
acquires reosurces from the resource module to avoid breaking these
tests.