build: ensure HWLOC_CFLAGS is set when building with hwloc
Problem: HWLOC_CFLAGS was missing (or incorrectly marked as
HWLOC_INCLUDES) in a couple of CPPFLAGS in the build. This caused
builds against libhwloc from alternate location to use the includes
from one version of hwloc, and the libs of another, causing potential
segfaults at runtime.

Fix the use of HWLOC_CFLAGS.