python: add prepend option to util.OutputFormat class
Problem: flux.util.OutputFormat class requires an override of its
get_format() method in order to prepend string to each format field.
(e.g. as in "0." to force fields to result in getattr() on a target
object) This will result in duplication of code if this feature
is used in more than one place.

Add a new, optional parameter `prepend` to the OutputFormat constructor,
and a move the code to reconstruct all format fields with a prepended
string into the class. If `prepend` is set, then the format string
will be replaced with a prepended version immediately.