terminus: fix compilation under gcc 10.1.0
Problem: gcc reports:
```
flux-terminus.c: In cmd_:
flux-terminus.c:710:15:  may be used uninitialized in this function
[-Werror=maybe-uninitialized]
  710 |     if (!(f = flux_rpc_pack (h, service, rank, 0,
      |               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  711 |                              "{s:i s:i s:i}",
      |                              ~~~~~~~~~~~~~~~~
  712 |                              "id", id,
      |                              ~~~~~~~~~
  713 |                              "signal", SIGKILL,
      |                              ~~~~~~~~~~~~~~~~~~
  714 |                              "wait", 1))
      |                              ~~~~~~~~~~
```
This is caused by the fact that `optparse_fatal_usage` is not guaranteed
to exit the program if the fatal function is overridden.  This means the
program could be free to continue running and read from the unintialized
`id` value.

Solution: Initialize `id` to -1.