libsubprocess: add functions to manipulate arguments
Problem: libsubprocess has functions to easily build the argv for a
subprocess, but there is no facility for manipulating an existing
argv. This functionality may be required for programs which need to
manipulate an existing flux_cmd_t, either as part of a plugin or
prepending wrappers, etc.

Add flux_cmd_argv_delete() and flux_cmd_argv_insert() to allow
arbitrary modification of an existing flux_cmd_t object.