flux-mini: add batch subcommand
Add a flux-mini "batch" subcommand which presents a minimal batch
submission interface for flux.

flux-mini batch is similar to the other flux-mini commands, but takes
a job script (and its arguments) on the commandline, and submits a
jobspec which will run this script as the initial program of a new
Flux instance.

As such, the job specification options for flux-mini batch are slightly
different than those of "run" and "submit". Since the parallel job
being submitted is a new Flux instance, the job specification of the
batch command deals in "task-slots" instead of number of tasks, e.g.

  -n, --nslots=N          Number of total slots requested
  -c, --cores-per-slot=N  Number of cores to allocate per slot
  -g, --gpus-per-slot=N   Number of GPUs to allocate per slot

flux-mini batch additionally presents an option to pass extra options
to the flux-brokers invoked when starting the instance:

  -B, --broker-opts=OPT   Pass options to flux brokers

When run, flux-mini batch ensures that its SCRIPT argument is indeed
a script via a heuristic (are the first two bytes "#!"). The script
is then copied into the jobspec under 'attributes.system.batch.script',
and the shell option 'per-resource' is set to 'type=node' such
that one flux-broker is executed per node reosurce by default, as
would be expected when launching a new child instance.