shell: affinity: remove flags from hwloc_topology_restrict()
Problem: The job shell affinity plugin calls hwloc_topology_restrict(3)
with various RESTRICT flags, but `flux-hwloc reload` does not use any
flags when restricting topology. This may cause a mismatch in logical
core ids between the affinity plugin and Flux proper. Resulting in errors
of the form

 ERROR: affinity: affinity: core1 not in topology
 WARN: affinity: unable to get cpuset for cores 1. Disabling affinity

presumably because the flags used by the affinity plugin are causing
the logical Core ids to be renumbered.

Remove the flags used in the affinity plugin, for consistency with the
loaded topology and core ids used in resource.hwloc.by_rank object.

Fixes #2965