flux-mini: add alloc subcommand
Add a flux-mini "alloc" subcommand which presents a minimal "allocate
new child instance" interface for Flux.

flux-mini alloc is similar to flux-mini batch, but is used to start
interactive child Flux instances. Thus, flux-mini alloc creates a
jobspec that executes

 flux broker [broker-opts]

by default. If a COMMAND is specified on the command line, then it is
used as the initial program of the new instance instead of the default
(the current shell).

The "pty" shell option is also set by default so that the job shell
allocates a pseudoterminal, allowing interactive use of the new
instance.

Once the job is submitted, flux-mini alloc behaves like flux-mini
run and invokes `flux job attach` on the JOBID.