flux-hwloc: add coreids list to generated by_rank object
Problem: The by_rank aggregate generated by flux-hwloc reload includes
the cpuset of all available cores, but does not include which ids in
that set belong to the actual core objects. This makes it impossible
to form a list of cores exclusive of PUs in users like sched-simple.

Add a "coreids" key to the by_rank object which includes only the
core ids. On systems with hyperthreads or multiple PUs per core, this
will be a subset of the cpuset.