flux-admin: add cleanup-push subcommand
Provide the ability to add commands to be run during
the new cleanup state.  Commands are pushed to the front
of the cleanup command list so, as with atexit(3), they
execute in the reverse order of registration.

A single command may be pushed with:
    flux admin cleanup-push COMMAND ...

Multiple commands may be pushed with:
    flux admin cleanup-push <<-EOT
    COMMAND ...
    COMMAND ...
    EOT

(one per line).  When a block of commands is pushed
in this way, the commands within the block execute
in the order they appear on stdin.

The main driver for this is to move commands out of
rc1 that must run before modules are unloaded or brokers
are taken offline, as discussed in #2909.

A secondary bonus is that instance owners could push
their own tasks to be run after their initial program
completes.  For example, if they want the instance to
wait for all enqueued work to complete, they could
run "flux admin cleanup-push flux queue idle".