testsuite: fix tests that use nproc(1)
Problem: Tests using output of nproc(1) as a surrogate for the
number of schedulable cores may fail on systems with hyperthreading,
since nproc reports number of processing units not the number of
cores, and sched-simple now schedules cores by default.

Update tests using nproc in this way to use the output of
flux-resource list instead, to get the actual number of available
cores.

For t2700-mini-cmd.t, reload sched-simple with sched-PUs to fall
back to previous behavior. This allows MULTICORE prereq to be satisfied
on systems with multiple HTs, but a single core.