python: add get() method to Future base class
Many of the underlying libflux calls that return a flux_future_t
do not have a future result, but are only error/no error. Thus,
it will be useful to have a Future.get() method that calls the
underlying flux_future_get() as

 flux_future_get(f, NULL)

to check if a call returning a Future has succeeded or failed.
Calls which return a Future that *does* contain a result should then
subclass Future and override the default get() method (e.g. RPC)