modules/pymod: use flux_reactor_run in echo.py
Pymod scripts will not be able to use the wrapped handle.reactor_run()
since this installs a signal watcher for SIGINT, and the broker will
already have signal watcher for this signal installed.

Access the reactor as handle.flux_reactor_run to ensure we use the
underlying ffi flux_reactor_run() instead of the wrapped version.