content-files: add "files" backing store demo
This is a demo content backing store module that stores
content in one big flat directory, using blobrefs as
file names (or other keys for checkpoint key-value pairs).

If the 'testing' module option is set, skip registering
as the cache's backing store so the module can be tested
in isolation.

Include standalone test programs that load and store directly
to the back-end file database.