content-cache: allow pre-set content.backing-module
Problem: need a way to select the content backing store
type for an instance.

Rename the 'content.backing' broker attribute that reflects
the current name of the registered backing store to
'content.backing-module'.  Allow this attribute to be
(optionally) set on the broker command line.  The attribute
can then be tested in rc1.

Once set, don't allow the backing store type to change,
since reloading the backing store module is OK, but unloading
one type and loading another type will result in data loss.