libcontent: factor out some reusable functions
The utility functions for registering/unregistering a content
back end, and the one for synchronously registering services
can be pulled out of content-sqlite.c for reuse in other
backing store modules.

Change these functions to log detailed error messages to
the broker on failure.

Also: the max blob size check in the backing store module is
redundant with the one in the content cache.  Guests are not
allowed to write directly to the backing store, so this limit
is not likely to be reached under normal circumstances.  Drop it
rather than promote it to a libcontent reusable.