shell: trap SIGALRM and forward to tasks
Problem: When a job hits its timelimit, the exec system sends SIGALRM
to the job shells, but the shell doesn't handle SIGALRM specifically and
thus is immediately terminated.

Add SIGARLM to the list of trapped signals in the shell, so that this
signal is instead forwarded on to job tasks.