job-exec: send SIGTERM on exception even if one is in progress
Send a SIGTERM to the job for each sev=0 job exception, even if
an exception is already in progress.

This allows a SIGTERM to be sent to a job that has timed out, or
otherwise has received a fatal job exception which didn't result in
SIGTERM in the first place.