t2233-job-info-security.t: fix update_job_userid()
Problem: The t2233-job-info-security.t test uses a function to
rewrite job userid in the eventlog for testing purposes. However,
if a userid set in an event is -1 (a valid entry for
FLUX_USERID_UNKNOWN), then a number with a trailing - is substitued
and the eventlog becomes corrupted.

Allow minus sign in userid pattern in the update_job_userid()
shell function.