sched-simple: fix unlikely allocation leak
Problem: In the unlikely case that an "R" string fails to be
generated in the simple scheduler, then a successful allocation
will could by leaked. That is the resources are allocated, but
the scheduler never sends an alloc response.

Send a "denied" response to the job-manager when Rstring_create()
fails, and be sure to revoke the successful allocation.