python: fix active refcount in h.reactor_run method
Problem: The default Flux handle reactor_run() method adds a
SignalWatcher to the reactor in order to make it interruptible,
but this adds an extra "active watcher" refcount to the reactor
which means the reactor will now never exit by default.

To preserve existing callers that depend on reactor_run() to exit
when all *their* active watchers have been stopped or destroyed,
decrease the refcount on the reactor before entering the *real*
flux_reactor_run(3). Increase it again before the SignalWatcher
is destroyed.

Fixes #3001