job-info: return useful error message in responses
Problem: job-info.list,list-id, and list-inactive return EINVAL
for a number of malformed requests, and it may be difficult for
a Flux user crafting these requests by hand (from Python for example)
to determine exactly why the request was invalid.

Add job_info_error_t structure that can be passed around between
job-info functions to fill in a textual error on failure. This error
can then be returned in the response payload, giving a specific
reason for the EINVAL and/or EPROTO response.

E.g. instead of generic EINVAL error, if an invalid attribute is
sent in the request, the list service now returns:

 badattr is not a valid attribute

or if jobid is not an integer

 message payload error: Expected integer, got string