python: attempt to encode correct types in job.job_list*
Problem: In the job.job_list* python bindings, message payloads
are created using whatever types were passed in by the caller. This
can unnecessarily cause EPROTO errors when, for example, a jobid
is passed as a string instead of an int.

Attempt to cast simple types to the expected payload type in the
job.job_list* functions. This way, users of the interface do not have
to worry about the underlying required types, and if a invalid parameter
is used, it will be caught before sending the message, not as an EPROTO
error from an RPC.