travis-ci: remove -k from make check
Problem: Addition of -k option to make in travis_run.sh causes
Travis-CI testing to continue even after errors, which lengthens the
time for failed builds, and makes it difficult to diagnose errors
early in the build process.

Remove -k from make, so that builds in Travis stop as soon as possible
on failure.