travis-ci: enforce correct configure ARGS for docker tags
Problem: When DOCKER_TAG=t is set for a Travis builder, the build
must have the correct corresponding configure ARGS, so that flux-core
is installed to prefix, built against flux-security, etc. O/w, the
docker image will fail to be installed properly.

Force ARGS to be correct when DOCKER_TAG is set, instead of setting
args for each individual builder.