docker: emit error if distcheck is used with some configure args
Problem: 'make distcheck' is not compatible with some configure args
that cause 'make install' to install somewhere outside of --prefix.
This is because distcheck attempts to overwrite prefix to point to
its special installation path, but args like --sysconfdir=/etc
or --with-systemdsystemunitdir explicitly try to install files outside
of this path, which fails.

Emit a fatal error when these args are used with -d, --distcheck in
docker-run-check.sh to prevent difficult to diagnose errors in the
future.