testsuite: fix LUA_PATH in t2603-job-shell-initrc.t
Problem: On some systems, the flux-shell initrc tests under
shell/initrc/tests may fail with an error finding the required
Lua posix module. This is because once the LUA_PATH is appended
to at the top of t2603-job-shell-initrc.t, the default paths are
overwritten (may be Lua version dependent).

Pick up the explicit built-in LUA_PATH from Lua itself and then
append to the path, instead of only using a (probably unset) existing
LUA_PATH.