docker: add fluxorama Dockerfile and build files
Add the current version of the "fluxorama" docker image source
Dockerfile and associated components to src/test/docker/fluxorama.

Fluxorama is a system instance docker image for flux-core built
on top of the flux-core:centos8 image. It runs a single broker
system instance via systemd, along with a ttyd server so that
a shell can be obtained by pointing a browser to the ttyd port
exposed by the container. The default user is 'fluxuser', with
the flux instance owner as user 'flux'. There are 5 other users
created in the container to facilitate multi-user testing.

To run:

 FLUXUSER_PASSWORD=mypass \
    docker run -ti \
    -e FLUXUSER_PASSWORD \
    --tmpfs=/run \
    -v /sys/fs/cgroup:/sys/fs/cgroup:ro \
    -p 8000:7681 \
    fluxrm/fluxorama

FLUXUSER_PASSWORD must be set or the container will refuse to be
run.

Then point a browser at localhost:8000 or run in the container with

 docker exec -u fluxuser -ti <container_id> bash

If "logging in" via the browser, supply the password set in the
FLUXUSER_PASSWORD environment variable.