broker/runat: add runat_is_defined, runat_is_completed
Problem: there is not an easy way for users of the runat
class to check whether a script list is defined without
running it.  It is also not possible to check whether a
script list has completed other than the completion callbacak.

Add the runat_is_defined() and runat_is_completed() accessors.

Update unit test.