libutil: fluid: add fluid_parse()
Problem: fluid_decode(3) requires a priori knowledge of the encoding
of the source string, and doesn't support integer representations of a
FLUID as decimal or hex. This makes usage of the function multi-step
for utilities that want to allow FLUID input from any representation.

Add a function to auto-detect the encoding type of a FLUID string
representation, along with a new `fluid_parse()` function that
decodes from any supported representation.