flux-job: id: use flux_job_id_parse/encode for conversion
Update the `flux-job id` utility to use the new flux_job_id_parse()
and flux_job_id_encode() functions for parsing JOBIDs from the command
line, and for conversion to the requested representation.

The --from=TYPE option is no longer required, since flux_job_id_parse
can unambiguously decode from all supported FLUID/JOBID encodings.
This option is dropped.

This change results in a renaming of the "hex" encoding to "dothex".
A "hex" representation is still supported, but is encoded as hexidecimal
with a "0x" prefix.