libjob: add flux_job_id_parse(), flux_job_id_encode()
Problem: FLUID string encoding is not exported via a public API,
but this functionality may be useful and/or necessary for callers,
especially for the case of converting JOBIDs between different
encodings.

Add flux_job_id_parse() and flux_jobid_encode() to libjob to allow
Flux API users to easily parse a Flux JOBID from any supported
string representation, or encode a `flux_jobid_t` to any supported
representation. The supported representations currently include
all standard FLUID encodings:

 - decimal integer
 - hexidecimal integer, prefixed with "0x"
 - dotted hex format, "XXXX.XXXX.XXXX.XXXX"
 - KVS path, "job.XXXX.XXXX.XXXX.XXXX"
 - Words, "X-X-X-X--X-X-X-X"
 - RFC 19 F58 encoding, prefixed with ƒ or f

Addition of these functions to the libflux API will also make it
easy to add support to bindings for encoding and decoding JOBIDs.

The functions are named `flux_job_id_*` instead of `flux_jobid_*`
to keep consistent naming with the other libjob functions, as well
as to make it easy to call the functions from Python cffi.