flux-jobs: force default stdout encoding to utf-8
Problem: Python3 doesn't work well in non UTF-8 clean locales
such as C/POSIX. This will be especially true when RFC19 F58
encoded JOBIDs are supported in flux-jobs.

Force stdout to UTF-8 encoding in flux-jobs. This is only necessary
in python 3.6 since UTF-8 seems to be the default in 3.7 and above.