flux-hwloc: remove ignore of `HWLOC_OBJ_GROUP`
Problem: When the `HWLOC_OBJ_GROUP` is ignored, the GPUs on the
Sierra/Lassen clusters are represented in the resource topology as
direct children of the node.  This topology ignores the fact that the
GPUs actually have locality with respect to the CPU sockets.  This
topology also is causing downstream affects with the fluxion scheduler
and its testsuite.  Depending on how the hwloc is read (either via
flux-core's flux-hwloc or directly via the hwloc API), the resource
topology changes (i.e., GPUs are children of the node versus the
sockets).  Also worth noting that the GPUs are children of the sockets
when using the hwloc V2 API, so ignoring the group creates a significant
difference in the topologies between hwloc versions.

Solution: remove the call to ignore `HWLOC_OBJ_GROUP` so that on the
Sierra/Lassen systems, the GPUs are children of the sockets.  This also
normalizes the resource topology across reading methods and hwloc
versions.  Now requesting a GPU on Sierra/Lassen can always be done with
a `node->socket->gpu` jobspec.

Related PR: flux-framework/flux-sched/pull/677