docker: ensure UTF8 locale set in all testenv images
Attempt to set a sane value for LANG environemnt variable in all
base "testenv" docker images, so that this is inherited by all
subsequently derived docker images.

Attempt to standardize on the POSIX locale, C.UTF-8. This seems
to be available in our all images except centos7. In that image
fall back to en_US.UTF-8.