t2800-jobs-cmd.t: refactor workload and tests
Problem: t2800-jobs-cmd.t hard codes the number of pending, running,
inactive, etc. jobs throughout the tests. Since the hard-coded values
often were without comment, it would difficult for a future developer
to add a job or modify the jobs "workload" without breaking tests,
and there is no way to search which tests might break due to a change.

Furthermore, the initial workload submission is slow because it
sumbits jobs serially, partially in an attempt to guarantee ordering.

Introduce a new system for "workload" submission (with copious
comments to document the purpose of the workload for future test
developers). Keep job id lists in ${state}.ids files, where state
can be one  of "active", "all", "cancelled", "completed", "failed",
"run", "sched", or "inactive". Add simple shell functions to query
these states, returning either the id list or a count of ids. Use
these new functions throughout the testsuite, include `wait_states`
synchronization.

While updating the testsuite, attempt to reduce the amount of
repetition, and especially try to reduce the redundant calls to
`flux-jobs`, since this is a python script and thus is fairly slow.