build: remove NEWS link, move COPYING to LICENSE
Problem: As stated in #3043, existence of LICENSE as a link to COPYING
is confusing GitHub project license detection. The presence of a COPYING
file is only required by automake adherence to strict GNU standards,
and now the project uses the 'foreign' option, so this file is
no longer necessary.

Move COPYING to LICENSE to satisfy GitHub.

Also remove NEWS, since it is no longer required as well.

Fixes #3043