build: do not include built manpages in distribution
Problem: doc/Makefile.am includes the built manpages in dist_man_MANS,
which causes the manpages to be built as result of `make dist` and also
included in the distribution tarball. This isn't a problem in itself,
but is not ideal since a) sphinx requirement for docs is still a
requirement for configure and b) man pages are cleaned durin `make clean`
and will have to be rebuilt anyway.

To avoid confusion, avoid including build products in distribution
tarball, and use `man_MANS` instead `dist_man_MANS`.

Fixes #3088