broker/state_machine: log signal handling errors
Problem: when a signal is processed by the state machine,
there is no logging if it cannot be translated into termination
of a runat script list.

Log those errors, and change the function signature to allow
the signal number to be shown in the log message.  Rename the
function to state_machine_kill() to reflect its slightly more
specialized role.