broker/state_machine: handle events reactively
Problem: when a state machine action posts an event, the new
action, if any, is called directly on the stack of the first action,
without anything else running in between.

Since it's generally a best practice to use the reactor to process
events, change the state machine to post events to a queue, and have
the queue execute reactively, using the prep/check/idle pattern.