flux-job: normalize output of jobids
Problem: flux-job subcommands are inconsistent when emitting jobids
in output and error log messages. Additionally, when error messages
reference jobids it is always by integer, when the user may have used
a different JOBID encoding on the commandline.

By default, output F58 encoded jobids from commands like flux-job
submit, and in other commands, be sure to save the "jobid" entered by
the user, and use that string when logging, instead of always using
the integer representation.