flux-mini: use JobID class when representing jobids
Problem: flux-mini subcommands always display integer jobid, whereas
the string representation of the JobID class should be used.

When a jobid is returned by a RPC and will be displayed to the user,
use it to construct a JobID object and use that object's string
representation when displaying the JobID to the user or in constructing
other commands.

Ensure that both stdout and stderr are using utf-8 encoding to avoid
encoding issues in Python <= 3.6.

Also ensure that when explictly casting JobID to string, as when
constructing arguments for os.exec*, that an explicit encoding is
used with `errors='surrogateescape'`. (This also works around encoding
errors on earlier versions of Python 3.6)