flux-job: exit with failure from flux-job id with no input
Problem: `flux job id` exits with success when no arguments or
input is received. It would be useful for testing if instead
flux-job id failed in this case, so that the construct

 flux job submit | flux job id

will fail even without something like `set -o pipefail` set
for the current shell.

Require at least one jobid to be processed to exit with success
in flux-job id.