libschedutil: preserve ordering of alloc requests
Problem: alloc requests may be delivered to the scheduler out of
order if asynchronous kvs operations are not completed in the same
order they are issued. This is because the scheduler alloc callback
is called as soon as the KVS lookup future is fulfilled, and there is
no guarantee about the order lookups are answered by the KVS. This
can cause resources to be allocated to a job with a lower priority
or later submission time, even with a FIFO schduler.

Add a queue for alloc requests, and use the queue order to deliver
alloc requests to the scheduler as they become ready, ensuring alloc
request ordering.

Fixes #3116