flux-mini: add errors="surrogateescape" to stdout/err
Problem: Even forcing the encoding of stdout and stderr to utf-8
is not enough to fully prevent Python from complaining when trying
to output utf-8 to the terminal.

Add errors="surrogateescape" as an additional arg when re-opening
stdout and stderr in utf-8 mode. This suppresses many of these
warnings and makes the script behave sensibly in most cases.