travis-ci: fix ARGS when DOCKER_TAG set
Problem: When DOCKER_TAG is set, ARGS is overwritten with the correct
installation configure arguments. This erases other required args
for the build including --enable-content-s3.

Be sure to keep existing ARGS when setting the DOCKER_TAG args.

This only affects merges, not PRs, which is why it wasn't discovered
by PR Travis checks.