cmd: attempt to set utf-8 locale in flux cmddriver
Problem: Use of non-UTF-8 locale may cause some inconveniences for
the casual Flux user due to the use of RFC19 F58 encoded JOBIDs
for most user-facing commands in Flux.

When setting up env in flux(1) command driver, check for a LC_ALL
setting that is empty, or set to "C" or "POSIX", and attempt to
"upgrade" to a common UTF-8 locale if so. Prefer C.UTF-8, but if
this fails, fall back to en_US.UTF-8.

If setlocale(3) on either of these locales fails, issue a warning
to the user.

Fixes #3124